Move Monogame Accessibility Singletons to separate folder
