+++ Classes/MonogameAccessibility/GraphicsDeviceManagerSingleton.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Xna.Framework.Graphics;
-namespace HandsOnDeck.Classes.Managers
+namespace HandsOnDeck.Classes.MonogameAccessibility
 {
     public class GraphicsDeviceSingleton
     {
