+++ Classes/UI/UIElements/UIElement.cs
@@ -1,4 +1,4 @@
-﻿using HandsOnDeck.Classes.Managers;
+﻿using HandsOnDeck.Classes.MonogameAccessibility;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
