+++ Game1.cs
@@ -3,6 +3,7 @@ using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 using System;
 using HandsOnDeck.Classes.Managers;
+using HandsOnDeck.Classes.MonogameAccessibility;
 
 namespace HandsOnDeck
 {
